pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "amplify"
version = "0.0.0"
description = "A transformer model for protein design"
requires-python = ">= 3.10"
dependencies = [
    "torch==2.2.*",
    "accelerate==0.27.*",
    "deepspeed==0.13.*",
    "xformers==0.0.24.*",
    "transformers==4.38.*",
    "datasets==2.17.*",
    "wandb==0.16.*",
    "hydra-core==1.3.*",
    "numpy==1.26.*",
]
authors = [
    { name = "Quentin Fournier", email = "quentin.fournier@mila.quebec" },
    { name = "Benjamin Schulz", email = "bschul02@amgen.com" },
    { name = "Robert Vernon", email = "rverno01@amgen.com" },
]
maintainers = [
    { name = "Quentin Fournier", email = "quentin.fournier@mila.quebec" },
    { name = "Benjamin Schulz", email = "bschul02@amgen.com" },
    { name = "Robert Vernon", email = "rverno01@amgen.com" },
]
readme = "README.md"
license = { text = "MIT License" }
keywords = ["Machine Learning", "Large Language Models", "Protein Language Models"]

[project.optional-dependencies]
dev = ["sphinx", "black", "piccolo-theme", "pytest", "pytest-cov", "pytest-clarity", "xeger"]
